+++ CHANGELOG.md
@@ -49,6 +49,7 @@ mode line indicator prefix customizable.
 * New defcustom `projectile-ignored-projects` allows you to specify projects
 that shouldn't be added to the known projects list.
 * New command `projectile-remove-current-project-from-known-projects`.
+* New defcustom `projectile-buffers-filter-function`.
 
 ### Changes